Correcting size of variable in fork and lfork

+++ corewar/src/instructions/instruction_lfork.c
@@ -10,7 +10,7 @@
 
 int instruction_lfork(vm_t *vm, champion_t *champion)
 {
-    ssize_t index = 0;
+    short index = 0;
     int tmp_pc = 0;
 
     if (!vm || !champion)
